+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to this project will be documented in this file.
 - *(ai)* Adapters and gemini ai working together
 - *(ai)* Wikipedia adapter and gemini ai working together
 - *(ai)* Input save at db or get from adapter working
+- *(question)* Generating/querying/saving question db/gemini implemented
 
 ### ⚙️ Miscellaneous Tasks